% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_filter}
\alias{add_filter}
\title{Add filter}
\usage{
add_filter(query, filter, base_url)
}
\arguments{
\item{query}{a query url.}

\item{filter}{a filter string.}

\item{base_url}{api base url.}
}
\description{
Add a filter to a query url.
}
