% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igdb_games.R
\name{igdb_games}
\alias{igdb_games}
\title{IGDB Games}
\usage{
igdb_games(search = NULL, id = NULL, n = 50, limit = 50,
  scroll = FALSE, filter = NULL, fields = NULL, order = NULL,
  api_key = igdb_key())
}
\arguments{
\item{search}{a string with a keyword or title.}

\item{id}{a vector containing game ids.}

\item{n}{maximum number of results to return.}

\item{limit}{maximum number of results per page.}

\item{scroll}{an alternative method for retrieving multiple pages of results.}

\item{filter}{a string that limits the kinds of results returned.}

\item{fields}{a vector of fields to retrieve. NULL returns a data.frame with
all available fields that contain data.}

\item{order}{a string defining how results are sorted.}

\item{api_key}{the IGDB api key.}
}
\description{
Search for a list of games from the IGDB API.
}
\examples{
# Search for the game Bioshock, ordering by the release date.
bioshock <- igdb_games(
  search = "bioshock",
  order = "first_release_date:asc"
)

}
\seealso{
\href{https://igdb.github.io/api/about/welcome/}{IGDB API}
}
