% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igdb_pulses.R
\name{igdb_pulses}
\alias{igdb_pulses}
\title{IGDB Pulses}
\usage{
igdb_pulses(search = NULL, id = NULL, n = 50, limit = 50,
  scroll = FALSE, filter = NULL, fields = NULL, order = NULL,
  api_key = igdb_key())
}
\arguments{
\item{search}{a string with a keyword or title.}

\item{id}{a vector containing game ids.}

\item{n}{maximum number of results to return.}

\item{limit}{maximum number of results per page.}

\item{scroll}{an alternative method for retrieving multiple pages of results.}

\item{filter}{a string that limits the kinds of results returned.}

\item{fields}{a vector of fields to retrieve. NULL returns a data.frame with
all available fields that contain data.}

\item{order}{a string defining how results are sorted.}

\item{api_key}{the IGDB api key.}
}
\description{
Search for a list of new articles from the IGDB API.
}
\examples{
\dontrun{
igdb_pulses(search = "activision")
}

}
\seealso{
\href{https://igdb.github.io/api/about/welcome/}{IGDB API}
}
